TST: Fix more tests and silence more warnings
[skip docs]

+++ tests/c/test_cchkbd.c
@@ -359,9 +359,9 @@ static void run_zchkbd_single(zchkbd_params_t* params)
     } else if (itype == 10) {
         f32 temp1 = -2.0f * logf(ulp);
         for (INT j = 0; j < mnmin; j++) {
-            BD[j] = exp(temp1 * rng_uniform_symmetric_f32(rng));
+            BD[j] = expf(temp1 * rng_uniform_symmetric_f32(rng));
             if (j < mnmin - 1)
-                BE[j] = exp(temp1 * rng_uniform_symmetric_f32(rng));
+                BE[j] = expf(temp1 * rng_uniform_symmetric_f32(rng));
         }
 
         iinfo = 0;
